Added server communication and execution loop to listen()
Listen now takes a username, password and opponent option as arguments

Edited makeMove() to accept an arraylist of moves, and make them. Needs
review to be sure that it functions as intended, as well as make sure
that it can move first as well. Easy as just taking out the null
stipulations?

Edited Player to allow for changing color of the player, since it's
determined by the server.

NOTE: I don't know if I got better or the AI got worse, but I can beat
it pretty easily